Add make_vocab.sh

+++ examples/swbd/s5c/run_3mtl.sh
@@ -8,6 +8,7 @@ echo "                              Switchboard                                 
 echo ============================================================================
 
 stage=0
+stop_stage=5
 gpu=
 speed_perturb=false
 specaug=false
@@ -27,18 +28,18 @@ vocab_sub2=300
 #########################
 # ASR configuration
 #########################
-conf=conf/asr/rnn_seq2seq_3mtl.yaml
+conf=conf/asr/blstm_las_3mtl.yaml
 conf2=
 asr_init=
 
 ### path to save the model
-model=/n/sd3/inaguma/result/swbd
+model=/n/work1/inaguma/results/swbd
 
 ### path to the model directory to resume training
 resume=
 
 ### path to save preproecssed data
-export data=/n/sd3/inaguma/corpus/swbd
+export data=/n/work1/inaguma/corpus/swbd
 
 ### path to original data
 SWBD_AUDIOPATH=/n/rd21/corpora_7/swb
@@ -103,7 +104,7 @@ if [ ${unit_sub2} != wp ]; then
     wp_type_sub2=
 fi
 
-if [ ${stage} -le 0 ] && [ ! -e ${data}/.done_stage_0 ]; then
+if [ ${stage} -le 0 ] && [ ${stop_stage} -ge 0 ] && [ ! -e ${data}/.done_stage_0 ]; then
     echo ============================================================================
     echo "                       Data Preparation (stage:0)                          "
     echo ============================================================================
@@ -134,30 +135,32 @@ if [ ${stage} -le 0 ] && [ ! -e ${data}/.done_stage_0 ]; then
     touch ${data}/.done_stage_0 && echo "Finish data preparation (stage: 0)."
 fi
 
-if [ ${stage} -le 1 ] && [ ! -e ${data}/.done_stage_1_${datasize}_sp${speed_perturb} ]; then
+if [ ${stage} -le 1 ] && [ ${stop_stage} -ge 1 ] && [ ! -e ${data}/.done_stage_1_${datasize}_sp${speed_perturb} ]; then
     echo ============================================================================
     echo "                    Feature extranction (stage:1)                          "
     echo ============================================================================
 
-    for x in train_swbd eval2000; do
-        steps/make_fbank.sh --nj 32 --cmd "$train_cmd" --write_utt2num_frames true \
-            ${data}/${x} ${data}/log/make_fbank/${x} ${data}/fbank || exit 1;
-        utils/fix_data_dir.sh ${data}/${x}
-    done
+    if [ ! -e ${data}/.done_stage_1_${datasize}_spfalse ]; then
+        for x in train_swbd eval2000; do
+            steps/make_fbank.sh --nj 32 --cmd "$train_cmd" --write_utt2num_frames true \
+                ${data}/${x} ${data}/log/make_fbank/${x} ${data}/fbank || exit 1;
+            utils/fix_data_dir.sh ${data}/${x}
+        done
 
-    # Use the first 4k sentences as dev set.
-    utils/subset_data_dir.sh --first ${data}/train_swbd 4000 ${data}/${dev_set} || exit 1;  # 5hr 6min
-    n=$[$(cat ${data}/train_swbd/segments | wc -l) - 4000]
-    utils/subset_data_dir.sh --last ${data}/train_swbd ${n} ${data}/${train_set}.tmp || exit 1;
+        # Use the first 4k sentences as dev set.
+        utils/subset_data_dir.sh --first ${data}/train_swbd 4000 ${data}/${dev_set} || exit 1;  # 5hr 6min
+        n=$[$(cat ${data}/train_swbd/segments | wc -l) - 4000]
+        utils/subset_data_dir.sh --last ${data}/train_swbd ${n} ${data}/${train_set}.tmp || exit 1;
 
-    # Finally, the full training set:
-    utils/data/remove_dup_utts.sh 300 ${data}/${train_set}.tmp ${data}/train_swbd_nodev || exit 1;  # 286hr
-    rm -rf ${data}/*.tmp
+        # Finally, the full training set:
+        utils/data/remove_dup_utts.sh 300 ${data}/${train_set}.tmp ${data}/train_nodev_swbd || exit 1;  # 286hr
+        rm -rf ${data}/*.tmp
 
-    if [ ${datasize} = fisher_swbd ]; then
-        steps/make_fbank.sh --nj 32 --cmd "$train_cmd" --write_utt2num_frames true \
-            ${data}/train_fisher ${data}/log/make_fbank/train_fisher ${data}/fbank || exit 1;
-        utils/combine_data.sh --extra_files "utt2num_frames" ${data}/${train_set} ${data}/train_swbd_nodev ${data}/train_fisher || exit 1;
+        if [ ${datasize} = fisher_swbd ]; then
+            steps/make_fbank.sh --nj 32 --cmd "$train_cmd" --write_utt2num_frames true \
+                ${data}/train_fisher ${data}/log/make_fbank/train_fisher ${data}/fbank || exit 1;
+            utils/combine_data.sh --extra_files "utt2num_frames" ${data}/${train_set} ${data}/train_nodev_swbd ${data}/train_fisher || exit 1;
+        fi
     fi
 
     if [ ${speed_perturb} = true ]; then
@@ -185,7 +188,7 @@ fi
 dict=${data}/dict/${train_set}_${unit}${wp_type}${vocab}.txt; mkdir -p ${data}/dict
 nlsyms=${data}/dict/nlsyms_${train_set}.txt
 wp_model=${data}/dict/${train_set}_${wp_type}${vocab}
-if [ ${stage} -le 2 ] && [ ! -e ${data}/.done_stage_2_${datasize}_${unit}${wp_type}${vocab}_sp${speed_perturb} ]; then
+if [ ${stage} -le 2 ] && [ ${stop_stage} -ge 2 ] && [ ! -e ${data}/.done_stage_2_${datasize}_${unit}${wp_type}${vocab}_sp${speed_perturb} ]; then
     echo ============================================================================
     echo "                      Dataset preparation (stage:2, main)                  "
     echo ============================================================================
@@ -194,29 +197,8 @@ if [ ${stage} -le 2 ] && [ ! -e ${data}/.done_stage_2_${datasize}_${unit}${wp_ty
     cut -f 2- -d " " ${data}/${train_set}/text | tr " " "\n" | sort | uniq | grep "\[" > ${nlsyms}
     cat ${nlsyms}
 
-    echo "Making a dictionary..."
-    echo "<unk> 1" > ${dict}  # <unk> must be 1, 0 will be used for "blank" in CTC
-    echo "<eos> 2" >> ${dict}  # <sos> and <eos> share the same index
-    echo "<pad> 3" >> ${dict}
-    [ ${unit} = char ] && echo "<space> 4" >> ${dict}
-    offset=$(cat ${dict} | wc -l)
-    if [ ${unit} = wp ]; then
-        if [ ${speed_perturb} = true ]; then
-            grep sp1.0 ${data}/${train_set}/text > ${data}/${train_set}/text.org
-            cp ${data}/${dev_set}/text ${data}/${dev_set}/text.org
-            cut -f 2- -d " " ${data}/${train_set}/text.org > ${data}/dict/input.txt
-        else
-            cut -f 2- -d " " ${data}/${train_set}/text > ${data}/dict/input.txt
-        fi
-        spm_train --user_defined_symbols=$(cat ${nlsyms} | tr "\n" ",") --input=${data}/dict/input.txt --vocab_size=${vocab} \
-            --model_type=${wp_type} --model_prefix=${wp_model} --input_sentence_size=100000000 --character_coverage=1.0
-        spm_encode --model=${wp_model}.model --output_format=piece < ${data}/dict/input.txt | tr ' ' '\n' | \
-            sort | uniq -c | sort -n -k1 -r | sed -e 's/^[ ]*//g' | cut -d " " -f 2 | grep -v '^\s*$' | awk -v offset=${offset} '{print $1 " " NR+offset}' >> ${dict}
-    else
-        text2dict.py ${data}/${train_set}/text --unit ${unit} --vocab ${vocab} --nlsyms ${nlsyms} --speed_perturb ${speed_perturb} | \
-            awk -v offset=${offset} '{print $0 " " NR+offset}' >> ${dict} || exit 1;
-    fi
-    echo "vocab size:" $(cat ${dict} | wc -l)
+    make_vocab.sh --unit ${unit} --nlsyms ${nlsyms} --wp_type ${wp_type} --wp_model ${wp_model} --character_coverage 1.0 --speed_perturb ${speed_perturb} \
+        ${data} ${dict} ${vocab} ${data}/${train_set}/text
 
     # normalize eval2000
     # 1) convert upper to lower
@@ -236,15 +218,11 @@ if [ ${stage} -le 2 ] && [ ! -e ${data}/.done_stage_2_${datasize}_${unit}${wp_ty
         mkdir -p ${data}/dict/word_count ${data}/dict/oov_rate
         echo "OOV rate:" > ${data}/dict/oov_rate/word${vocab}_${datasize}.txt
         for x in ${train_set} ${dev_set}; do
-            if [ ${speed_perturb} = true ]; then
-                cut -f 2- -d " " ${data}/${x}/text.org | tr " " "\n" | sort | uniq -c | sort -n -k1 -r \
-                    > ${data}/dict/word_count/${x}_${datasize}.txt || exit 1;
-            else
-                cut -f 2- -d " " ${data}/${x}/text | tr " " "\n" | sort | uniq -c | sort -n -k1 -r \
-                    > ${data}/dict/word_count/${x}_${datasize}.txt || exit 1;
-            fi
+            cut -f 2- -d " " ${data}/${x}/text | tr " " "\n" | sort | uniq -c | sort -n -k1 -r \
+                > ${data}/dict/word_count/${x}_${datasize}.txt || exit 1;
             compute_oov_rate.py ${data}/dict/word_count/${x}_${datasize}.txt ${dict} ${x} \
                 >> ${data}/dict/oov_rate/word${vocab}_${datasize}.txt || exit 1;
+            # NOTE: speed perturbation is not considered
         done
         for set in "swbd" "ch"; do
             cut -f 2- -d " " ${data}/${test_set}/text.${set} | tr " " "\n" | sort | uniq -c | sort -n -k1 -r \
@@ -271,57 +249,21 @@ fi
 # sub1
 dict_sub1=${data}/dict/${train_set}_${unit_sub1}${wp_type_sub1}${vocab_sub1}.txt
 wp_model_sub1=${data}/dict/${train_set}_${wp_type_sub1}${vocab_sub1}
-if [ ${stage} -le 2 ] && [ ! -e ${data}/.done_stage_2_${datasize}_${unit_sub1}${wp_type_sub1}${vocab_sub1}_sp${speed_perturb} ]; then
+if [ ${stage} -le 2 ] && [ ${stop_stage} -ge 2 ] && [ ! -e ${data}/.done_stage_2_${datasize}_${unit_sub1}${wp_type_sub1}${vocab_sub1}_sp${speed_perturb} ]; then
     echo ============================================================================
     echo "                      Dataset preparation (stage:2, sub1)                  "
     echo ============================================================================
 
-    echo "Making a dictionary..."
-    echo "<unk> 1" > ${dict_sub1}  # <unk> must be 1, 0 will be used for "blank" in CTC
-    echo "<eos> 2" >> ${dict_sub1}  # <sos> and <eos> share the same index
-    echo "<pad> 3" >> ${dict_sub1}
-    [ ${unit_sub1} = char ] && echo "<space> 4" >> ${dict_sub1}
-    offset=$(cat ${dict_sub1} | wc -l)
-    if [ ${unit_sub1} = wp ]; then
-        if [ ${speed_perturb} = true ]; then
-            grep sp1.0 ${data}/${train_set}/text > ${data}/${train_set}/text.org
-            cp ${data}/${dev_set}/text ${data}/${dev_set}/text.org
-            cut -f 2- -d " " ${data}/${train_set}/text.org > ${data}/dict/input.txt
-        else
-            cut -f 2- -d " " ${data}/${train_set}/text > ${data}/dict/input.txt
-        fi
-        spm_train --user_defined_symbols=$(cat ${nlsyms} | tr "\n" ",") --input=${data}/dict/input.txt --vocab_size=${vocab_sub1} \
-            --model_type=${wp_type_sub1} --model_prefix=${wp_model_sub1} --input_sentence_size=100000000 --character_coverage=1.0
-        spm_encode --model=${wp_model_sub1}.model --output_format=piece < ${data}/dict/input.txt | tr ' ' '\n' | \
-            sort | uniq -c | sort -n -k1 -r | sed -e 's/^[ ]*//g' | cut -d " " -f 2 | grep -v '^\s*$' | awk -v offset=${offset} '{print $1 " " NR+offset}' >> ${dict_sub1}
-    elif [ ${unit_sub1} = phone ]; then
-        map_lexicon.sh ${data}/${train_set} ${data}/local/dict_nosp/lexicon.txt > ${data}/${train_set}/text.phone
-        map_lexicon.sh ${data}/${dev_set} ${data}/local/dict_nosp/lexicon.txt > ${data}/${dev_set}/text.phone
-        text2dict.py ${data}/${train_set}/text.phone --unit ${unit_sub1} --nlsyms ${nlsyms} --speed_perturb ${speed_perturb} | \
-            awk -v offset=${offset} '{print $0 " " NR+offset}' >> ${dict_sub1} || exit 1;
-    else
-        text2dict.py ${data}/${train_set}/text --unit ${unit_sub1} --vocab ${vocab_sub1} --nlsyms ${nlsyms} --speed_perturb ${speed_perturb} | \
-            awk -v offset=${offset} '{print $0 " " NR+offset}' >> ${dict_sub1} || exit 1;
-    fi
-    echo "vocab size:" $(cat ${dict_sub1} | wc -l)
+    make_vocab.sh --unit ${unit_sub1} --nlsyms ${nlsyms} --wp_type ${wp_type_sub1} --wp_model ${wp_model_sub1} --character_coverage 0.9995 --speed_perturb ${speed_perturb} \
+        ${data} ${dict_sub1} ${vocab_sub1} ${data}/${train_set}/text
 
     echo "Making dataset tsv files for ASR ..."
-    if [ ${unit_sub1} = phone ]; then
-        make_dataset.sh --feat ${dump_dir}/feats.scp --unit ${unit_sub1} --text ${data}/${train_set}/text.phone \
-            ${data}/${train_set} ${dict_sub1} > ${data}/dataset/${train_set}_${unit_sub1}${wp_type_sub1}${vocab_sub1}.tsv || exit 1;
-    else
-        make_dataset.sh --feat ${dump_dir}/feats.scp --unit ${unit_sub1} --nlsyms ${nlsyms} --wp_model ${wp_model_sub1} \
-            ${data}/${train_set} ${dict_sub1} > ${data}/dataset/${train_set}_${unit_sub1}${wp_type_sub1}${vocab_sub1}.tsv || exit 1;
-    fi
+    make_dataset.sh --feat ${data}/dump/${train_set}/feats.scp --unit ${unit_sub1} --nlsyms ${nlsyms} --wp_model ${wp_model_sub1} \
+        ${data}/${train_set} ${dict_sub1} > ${data}/dataset/${train_set}_${unit_sub1}${wp_type_sub1}${vocab_sub1}.tsv || exit 1;
     for x in ${dev_set} ${test_set}; do
         dump_dir=${data}/dump/${x}_${datasize}
-        if [ ${unit_sub1} = phone ] && [ ${x} != ${test_set} ]; then
-            make_dataset.sh --feat ${dump_dir}/feats.scp --unit ${unit_sub1} --nlsyms ${nlsyms} --text ${data}/${x}/text.phone  \
-                ${data}/${x} ${dict} > ${data}/dataset/${x}_${datasize}_${unit_sub1}${wp_type_sub1}${vocab_sub1}.tsv || exit 1;
-        else
-            make_dataset.sh --feat ${dump_dir}/feats.scp --unit ${unit_sub1} --nlsyms ${nlsyms} --wp_model ${wp_model_sub1} \
-                ${data}/${x} ${dict} > ${data}/dataset/${x}_${datasize}_${unit_sub1}${wp_type_sub1}${vocab_sub1}.tsv || exit 1;
-        fi
+        make_dataset.sh --feat ${dump_dir}/feats.scp --unit ${unit_sub1} --nlsyms ${nlsyms} --wp_model ${wp_model_sub1} \
+            ${data}/${x} ${dict_sub1} > ${data}/dataset/${x}_${datasize}_${unit_sub1}${wp_type_sub1}${vocab_sub1}.tsv || exit 1;
     done
 
     touch ${data}/.done_stage_2_${datasize}_${unit_sub1}${wp_type_sub1}${vocab_sub1}_sp${speed_perturb} && echo "Finish creating dataset for ASR (stage: 2)."
@@ -335,35 +277,27 @@ if [ ${stage} -le 2 ] && [ ! -e ${data}/.done_stage_2_${datasize}_${unit_sub2}${
     echo "                      Dataset preparation (stage:2, sub2)                  "
     echo ============================================================================
 
-    echo "Making a dictionary..."
-    echo "<unk> 1" > ${dict_sub2}  # <unk> must be 1, 0 will be used for "blank" in CTC
-    echo "<eos> 2" >> ${dict_sub2}  # <sos> and <eos> share the same index
-    echo "<pad> 3" >> ${dict_sub2}
-    [ ${unit_sub2} = char ] && echo "<space> 4" >> ${dict_sub2}
-    offset=$(cat ${dict_sub2} | wc -l)
-    if [ ${unit_sub2} = wp ]; then
-        cut -f 2- -d " " ${data}/${train_set}/text > ${data}/dict/input.txt
-        spm_train --user_defined_symbols=$(cat ${nlsyms} | tr "\n" ",") --input=${data}/dict/input.txt --vocab=${vocab_sub2} \
-            --model_type=${wp_type_sub2} --model_prefix=${wp_model_sub2} --input_sentence_size=100000000 --character_coverage=1.0
-        spm_encode --model=${wp_model_sub2}.model --output_format=piece < ${data}/dict/input.txt | tr ' ' '\n' | \
-            sort | uniq -c | sort -n -k1 -r | sed -e 's/^[ ]*//g' | cut -d " " -f 2 | grep -v '^\s*$' | awk -v offset=${offset} '{print $1 " " NR+offset}' >> ${dict_sub2}
-    elif [ ${unit_sub2} = phone ]; then
+    if [ ${unit_sub2} = phone ]; then
+        echo "Making a dictionary..."
+        echo "<unk> 1" > ${dict_sub2}  # <unk> must be 1, 0 will be used for "blank" in CTC
+        echo "<eos> 2" >> ${dict_sub2}  # <sos> and <eos> share the same index
+        echo "<pad> 3" >> ${dict_sub2}
+        offset=$(cat ${dict_sub2} | wc -l)
         map_lexicon.sh ${data}/${train_set} ${data}/local/dict_nosp/lexicon.txt
         map_lexicon.sh ${data}/${dev_set} ${data}/local/dict_nosp/lexicon.txt
         text2dict.py ${data}/${train_set}/text.phone --unit ${unit_sub2} --nlsyms ${nlsyms} --speed_perturb ${speed_perturb} | \
             awk -v offset=${offset} '{print $0 " " NR+offset}' >> ${dict_sub2} || exit 1;
     else
-        text2dict.py ${data}/${train_set}/text --unit ${unit_sub2} --vocab ${vocab_sub2} --nlsyms ${nlsyms} --speed_perturb ${speed_perturb} | \
-            awk -v offset=${offset} '{print $0 " " NR+offset}' >> ${dict_sub2} || exit 1;
+        make_vocab.sh --unit ${unit_sub2} --nlsyms ${nlsyms} --wp_type ${wp_type_sub2} --wp_model ${wp_model_sub2} --character_coverage 0.9995 --speed_perturb ${speed_perturb} \
+            ${data} ${dict_sub2} ${vocab_sub2} ${data}/${train_set}/text
     fi
-    echo "vocab size:" $(cat ${dict_sub2} | wc -l)
 
     echo "Making dataset tsv files for ASR ..."
     if [ ${unit_sub2} = phone ]; then
-        make_dataset.sh --feat ${dump_dir}/feats.scp --unit ${unit_sub2} --text ${data}/${train_set}/text.phone \
+        make_dataset.sh --feat ${data}/dump/${train_set}/feats.scp --unit ${unit_sub2} --text ${data}/${train_set}/text.phone \
             ${data}/${train_set} ${dict_sub2} > ${data}/dataset/${train_set}_${unit_sub2}${wp_type_sub2}${vocab_sub2}.tsv || exit 1;
     else
-        make_dataset.sh --feat ${dump_dir}/feats.scp --unit ${unit_sub2} --nlsyms ${nlsyms} --wp_model ${wp_model_sub2} \
+        make_dataset.sh --feat ${data}/dump/${train_set}/feats.scp --unit ${unit_sub2} --nlsyms ${nlsyms} --wp_model ${wp_model_sub2} \
             ${data}/${train_set} ${dict_sub2} > ${data}/dataset/${train_set}_${unit_sub2}${wp_type_sub2}${vocab_sub2}.tsv || exit 1;
     fi
     for x in ${dev_set} ${test_set}; do
@@ -381,7 +315,7 @@ if [ ${stage} -le 2 ] && [ ! -e ${data}/.done_stage_2_${datasize}_${unit_sub2}${
 fi
 
 mkdir -p ${model}
-if [ ${stage} -le 4 ]; then
+if [ ${stage} -le 4 ] && [ ${stop_stage} -ge 4 ]; then
     echo ============================================================================
     echo "                       ASR Training stage (stage:4)                        "
     echo ============================================================================
